# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Agila API
  version: "1.0.0"
  description: >
    Agila API v1. Send Authorization Bearer with a session JWT or an ek_ token.
    The same handlers and permission checks as the web app apply.
    Relay routes add external keys, column titles, claim, and plans.
    OAuth callbacks are not part of this API.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  "/api/v1/boards":
    get:
      summary: "GET /api/v1/boards"
      description: "Implemented in server/routes/boards.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
    post:
      summary: "POST /api/v1/boards"
      description: "Implemented in server/routes/boards.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/boards/{id}/claim":
    post:
      summary: "POST /api/v1/boards/{id}/claim"
      description: "Implemented in server/routes/apiV1Relay.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/plans":
    post:
      summary: "POST /api/v1/plans"
      description: "Implemented in server/routes/apiV1Relay.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/sprints":
    post:
      summary: "POST /api/v1/sprints"
      description: "Implemented in server/routes/apiV1Relay.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/tasks":
    get:
      summary: "GET /api/v1/tasks"
      description: "Implemented in server/routes/tasks.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
    post:
      summary: "POST /api/v1/tasks"
      description: "Implemented in server/routes/tasks.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/tasks/{id}/move":
    post:
      summary: "POST /api/v1/tasks/{id}/move"
      description: "Implemented in server/routes/apiV1Relay.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/tasks/{taskId}/acceptance-criteria/{criterionId}":
    patch:
      summary: "PATCH /api/v1/tasks/{taskId}/acceptance-criteria/{criterionId}"
      description: "Implemented in server/routes/apiV1Relay.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
  "/api/v1/tasks/{taskId}/acceptance-criteria/{id}":
    delete:
      summary: "DELETE /api/v1/tasks/{taskId}/acceptance-criteria/{id}"
      description: "Implemented in server/routes/acceptanceCriteria.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
    put:
      summary: "PUT /api/v1/tasks/{taskId}/acceptance-criteria/{id}"
      description: "Implemented in server/routes/acceptanceCriteria.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...